Refactor client versions
