+++ app/clients/__init__.py
@@ -1,6 +1,6 @@
 
-from .b20130606 import ResponsePacket as DefaultResponsePacket
-from .b20130606 import RequestPacket as DefaultRequestPacket
+from .b20130815 import ResponsePacket as DefaultResponsePacket
+from .b20130815 import RequestPacket as DefaultRequestPacket
 
 from .reader import BaseReader
 from .writer import BaseWriter
@@ -8,7 +8,7 @@ from .writer import BaseWriter
 from . import packets
 from . import handler
 
-from . import b20130606
+from . import b20130815
 from . import b20130329
 from . import b20121223
 from . import b20121119
