+++ app/clients/b20130329/__init__.py
@@ -7,9 +7,9 @@ from .. import (
 from ..packets import PACKETS
 from copy import deepcopy
 
-# Inherit packets from 20130606
-PACKETS[20130329] = deepcopy(PACKETS[20130606])
-PACKETS[20130303] = deepcopy(PACKETS[20130606])
+# Inherit packets from 20130815
+PACKETS[20130329] = deepcopy(PACKETS[20130815])
+PACKETS[20130118] = deepcopy(PACKETS[20130815])
 
 from .writer import Writer
 from .reader import Reader
