+++ app/clients/b20130329/decoder.py
@@ -10,7 +10,7 @@ from typing import Callable
 def register(packet: RequestPacket) -> Callable:
     def wrapper(func) -> Callable:
         PACKETS[20130329][0][packet] = func
-        PACKETS[20130303][0][packet] = func
+        PACKETS[20130118][0][packet] = func
         return func
 
     return wrapper
