+++ app/clients/b20130329/encoder.py
@@ -10,7 +10,7 @@ from typing import Callable
 def register(packet: ResponsePacket) -> Callable:
     def wrapper(func) -> Callable:
         PACKETS[20130329][1][packet] = func
-        PACKETS[20130303][1][packet] = func
+        PACKETS[20130118][1][packet] = func
         return func
 
     return wrapper
