+++ app/clients/b20130815/__init__.py
@@ -9,7 +9,7 @@ from . import encoder
 
 from ..packets import PACKETS
 
-PACKETS[20130606][2] = RequestPacket
-PACKETS[20130606][3] = ResponsePacket
-PACKETS[20130418][2] = RequestPacket
-PACKETS[20130418][3] = ResponsePacket
+PACKETS[20130815][2] = RequestPacket
+PACKETS[20130815][3] = ResponsePacket
+PACKETS[20130401][2] = RequestPacket
+PACKETS[20130401][3] = ResponsePacket
